// src/app/enums/attraction-tags.ts
export enum AttractionTags {
  Art = "attraction.category.Art",
  Children = "attraction.category.Children",
  Christmas = "attraction.category.ChristmasTime",
  Conferences = "attraction.category.Conferences",
  Dance = "attraction.category.Dance",
  Education = "attraction.category.Education",
  Exhibitions = "attraction.category.Exhibitions",
  Festivals = "attraction.category.Festivals",
  Health = "attraction.category.Health",
  InformationEvents = "attraction.category.InformationEvents",
  Lectures = "attraction.category.Lectures",
  Music = "attraction.category.Music",
  Police = "attraction.category.Police",
  Politics = "attraction.category.Politics",
  Recreation = "attraction.category.Recreation",
  Seniors = "attraction.category.Seniors",
  Stages = "attraction.category.Stages",
  Walks = "attraction.category.Walks",
  WeeklyMarkets = "attraction.category.WeeklyMarkets",
  Women = "attraction.category.Women",
}